+++ quantify_scheduler/json_utils.py
@@ -26,12 +26,10 @@ class JSONSchemaValMixin:  # pylint: disable=too-few-public-methods
     @classmethod
     def is_valid(cls, object_to_be_validated) -> bool:
         """Checks if the object is valid according to its schema."""
-        # schema_filename = "schedule.json"
 
         scheme = load_json_schema(__file__, cls.schema_filename)
         jsonschema.validate(object_to_be_validated.data, scheme)
-        # _ = object_to_be_validated.hash  # test that the hash property evaluates
-        return True  # if not exception was raised during validation
+        return True  # if no exception was raised during validation
 
 
 class ScheduleJSONDecoder(json.JSONDecoder):
@@ -95,7 +93,7 @@ class ScheduleJSONDecoder(json.JSONDecoder):
 
         Parameters
         ----------
-        obj :
+        obj
             The dictionary to deserialize.
 
         Returns
@@ -123,7 +121,7 @@ class ScheduleJSONDecoder(json.JSONDecoder):
 
         Parameters
         ----------
-        obj : str
+        obj
             The value of dictionary pair to deserialize.
 
         Returns
@@ -158,7 +156,7 @@ class ScheduleJSONDecoder(json.JSONDecoder):
 
         Parameters
         ----------
-        obj :
+        obj
             A pair of JSON objects.
 
         Returns