+++ quantify_scheduler/types.py
@@ -13,10 +13,12 @@ from copy import deepcopy
 from dataclasses import dataclass
 from enum import Enum
 from pydoc import locate
-from typing import TYPE_CHECKING, Any, Dict, List, Type
+from typing import TYPE_CHECKING, Any, Dict, List, Optional, Tuple, Type, Union
 from uuid import uuid4
 
 import numpy as np
+from matplotlib.axes import Axes
+from matplotlib.figure import Figure
 from quantify_core.utilities import general
 from typing_extensions import Literal
 
@@ -91,7 +93,7 @@ class Operation(JSONSchemaValMixin, UserDict):  # pylint: disable=too-many-ances
 
         Parameters
         ----------
-        other :
+        other
 
         Returns
         -------
@@ -194,7 +196,7 @@ class Operation(JSONSchemaValMixin, UserDict):  # pylint: disable=too-many-ances
 
             Parameters
             ----------
-            key :
+            key
 
             Returns
             -------
@@ -218,7 +220,7 @@ class Operation(JSONSchemaValMixin, UserDict):  # pylint: disable=too-many-ances
 
         Parameters
         ----------
-        gate_operation :
+        gate_operation
             an operation containing gate_info.
         """
         self.data["gate_info"].update(gate_operation.data["gate_info"])
@@ -229,7 +231,7 @@ class Operation(JSONSchemaValMixin, UserDict):  # pylint: disable=too-many-ances
 
         Parameters
         ----------
-        pulse_operation :
+        pulse_operation
             an operation containing pulse_info.
         """
         self.data["pulse_info"] += pulse_operation.data["pulse_info"]
@@ -241,7 +243,7 @@ class Operation(JSONSchemaValMixin, UserDict):  # pylint: disable=too-many-ances
 
         Parameters
         ----------
-        acquisition_operation :
+        acquisition_operation
             an operation containing acquisition_info.
         """
         self.data["acquisition_info"] += acquisition_operation.data["acquisition_info"]
@@ -363,33 +365,33 @@ class Operation(JSONSchemaValMixin, UserDict):  # pylint: disable=too-many-ances
 # pylint: disable=too-many-ancestors
 class ScheduleBase(JSONSchemaValMixin, UserDict, ABC):
     """
-    The :class:`~quantify_scheduler.types.ScheduleBase` is a data structure that is at
+    The :class:`~.ScheduleBase` is a data structure that is at
     the core of the Quantify-scheduler and describes when what operations are applied
     where.
 
-    The :class:`~quantify_scheduler.types.ScheduleBase` is a collection of
-    :class:`~Operation` objects and timing constraints that define relations between
+    The :class:`~.ScheduleBase` is a collection of
+    :class:`~.Operation` objects and timing constraints that define relations between
     the operations.
 
     The schedule data structure is based on a dictionary.
     This dictionary contains:
 
-    - operation_dict - a hash table containing the unique :class:`~Operation` s added
+    - operation_dict - a hash table containing the unique :class:`~.Operation` s added
         to the schedule.
     - timing_constraints - a list of all timing constraints added between operations.
 
-    The :class:`~quantify_scheduler.types.Schedule` provides an API to create schedules.
-    The :class:`~quantify_scheduler.types.CompiledSchedule` represents a schedule after
+    The :class:`~.Schedule` provides an API to create schedules.
+    The :class:`~.CompiledSchedule` represents a schedule after
     it has been compiled for execution on a backend.
 
 
-    The :class:`~quantify_scheduler.types.Schedule` contains information on the
-    :attr:`~quantify_scheduler.types.ScheduleBase.operations` and
-    :attr:`~quantify_scheduler.types.ScheduleBase.timing_constraints`.
-    The :attr:`~quantify_scheduler.types.ScheduleBase.operations` is a dictionary of all
+    The :class:`~.Schedule` contains information on the
+    :attr:`~.ScheduleBase.operations` and
+    :attr:`~.ScheduleBase.timing_constraints`.
+    The :attr:`~.ScheduleBase.operations` is a dictionary of all
     unique operations used in the schedule and contain the information on *what*
     operation to apply *where*.
-    The :attr:`~quantify_scheduler.types.ScheduleBase.timing_constraints` is a list of
+    The :attr:`~.ScheduleBase.timing_constraints` is a list of
     dictionaries describing timing constraints between operations, i.e. when to apply
     an operation.
 
@@ -428,8 +430,8 @@ class ScheduleBase(JSONSchemaValMixin, UserDict, ABC):
         A dictionary of all unique operations used in the schedule.
         This specifies information on *what* operation to apply *where*.
 
-        The keys correspond to the :attr:`~Operation.hash` and values are instances
-        of :class:`~Operation`.
+        The keys correspond to the :attr:`~.Operation.hash` and values are instances
+        of :class:`~.Operation`.
         """
         return self.data["operation_dict"]
 
@@ -461,7 +463,7 @@ class ScheduleBase(JSONSchemaValMixin, UserDict, ABC):
         .. note::
 
             timing constraints are not intended to be modified directly.
-            Instead use the :meth:`~quantify_scheduler.types.Schedule.add`
+            Instead use the :meth:`~.Schedule.add`
 
         """
         return self.data["timing_constraints"]
@@ -499,7 +501,7 @@ class ScheduleBase(JSONSchemaValMixin, UserDict, ABC):
 
         Parameters
         ----------
-        data :
+        data
             The JSON data.
 
         Returns
@@ -512,17 +514,102 @@ class ScheduleBase(JSONSchemaValMixin, UserDict, ABC):
 
         return Schedule(name, data=schedule_data)
 
+    def plot_circuit_diagram_mpl(
+        self, figsize: Tuple[int, int] = None, ax: Optional[Axes] = None
+    ) -> Tuple[Figure, Union[Axes, List[Axes]]]:
+        """
+        Creates a circuit diagram visualization of the schedule using matplotlib.
+
+        The circuit diagram visualization visualizes the schedule at the quantum circuit
+        layer.
+        This visualization provides no timing information, only showing the order of
+        operations.
+        Because quantify-scheduler uses a hybrid gate-pulse paradigm, operations for
+        which no information is specified at the gate level are visualized using an
+        icon (e.g., a stylized wavy pulse) depending on the information specified at
+        the quantum device layer.
+
+        Alias of :func:`.circuit_diagram.circuit_diagram_matplotlib`.
+
+        Parameters
+        ----------
+        schedule
+            the schedule to render.
+        figsize
+            matplotlib figsize.
+        ax
+            Axis handle to use for plotting.
+
+        Returns
+        -------
+        fig
+            matplotlib figure object.
+        ax
+            matplotlib axis object.
+        """
+        # NB imported here to avoid circular import
+        # pylint: disable=import-outside-toplevel
+        import quantify_scheduler.visualization.circuit_diagram as cd
+
+        return cd.circuit_diagram_matplotlib(schedule=self, figsize=figsize, ax=ax)
+
+    # pylint: disable=too-many-arguments
+    def plot_pulse_diagram_mpl(
+        self,
+        port_list: Optional[List[str]] = None,
+        sampling_rate: float = 1e9,
+        modulation: Literal["off", "if", "clock"] = "off",
+        modulation_if: float = 0.0,
+        ax: Optional[Axes] = None,
+    ) -> Tuple[Figure, Axes]:
+        """
+        Creates a visualization of all the pulses in a schedule using matplotlib.
+
+        The pulse diagram visualizes the schedule at the quantum device layer.
+        For this visualization to work, all operations need to have the information
+        present (e.g., pulse info) to represent these on the quantum-circuit level and
+        requires the absolute timing to have been determined.
+        This information is typically added when the quantum-device level compilation is
+        performed.
+
+        Alias of :func:`.pulse_diagram.pulse_diagram_matplotlib`.
+
+        port_list :
+            A list of ports to show. if set to `None` will use the first
+            8 ports it encounters in the sequence.
+        modulation :
+            Determines if modulation is included in the visualization.
+        modulation_if :
+            Modulation frequency used when modulation is set to "if".
+        sampling_rate :
+            The time resolution used to sample the schedule in Hz.
+        ax:
+            Axis onto which to plot.
+        """
+        # NB imported here to avoid circular import
+        # pylint: disable=import-outside-toplevel
+        import quantify_scheduler.visualization.pulse_diagram as pd
+
+        return pd.pulse_diagram_matplotlib(
+            schedule=self,
+            sampling_rate=sampling_rate,
+            ax=ax,
+            port_list=port_list,
+            modulation=modulation,
+            modulation_if=modulation_if,
+        )
+
 
 class Schedule(ScheduleBase):  # pylint: disable=too-many-ancestors
     """
     A modifiable schedule.
 
-    Operations :class:`~quantify_scheduler.types.Operation` can be added using the
-    :meth:`~quantify_scheduler.types.Schedule.add` method, allowing precise
+    Operations :class:`~.Operation` can be added using the
+    :meth:`~.Schedule.add` method, allowing precise
     specification *when* to perform an operation using timing constraints.
 
     When adding an operation, it is not required to specify how to represent this
-    :class:`~quantify_scheduler.types.Operation` on all layers.
+    :class:`~.Operation` on all layers.
     Instead, this information can be added later during
     :ref:`compilation <sec-compilation>`.
     This allows the user to effortlessly mix the gate- and pulse-level descriptions as
@@ -538,16 +625,12 @@ class Schedule(ScheduleBase):  # pylint: disable=too-many-ancestors
 
         Parameters
         ----------
-        name :
+        name
             The name of the schedule
-        repetitions :
+        repetitions
             The amount of times the schedule will be repeated, by default 1
-        data :
+        data
             A dictionary containing a pre-existing schedule., by default None
-
-        Raises
-        ------
-        NotImplementedError
         """
 
         # validate the input data to ensure it is valid schedule data
@@ -606,26 +689,26 @@ class Schedule(ScheduleBase):  # pylint: disable=too-many-ancestors
         the operations.
         The reference operation (:code:`"ref_op"`) is specified using its label
         property.
-        See also :attr:`~quantify_scheduler.types.ScheduleBase.timing_constraints`.
+        See also :attr:`~.ScheduleBase.timing_constraints`.
 
         Parameters
         ----------
-        operation :
+        operation
             The operation to add to the schedule
-        rel_time :
+        rel_time
             relative time between the reference operation and the added operation.
             the time is the time between the "ref_pt" in the reference operation and
             "ref_pt_new" of the operation that is added.
-        ref_op :
+        ref_op
             label of the reference operation. If set to :code:`None`, will default
             to the last added operation.
-        ref_pt :
+        ref_pt
             reference point in reference operation must be one of
             ('start', 'center', 'end').
-        ref_pt_new :
+        ref_pt_new
             reference point in added operation must be one of
             ('start', 'center', 'end').
-        label :
+        label
             a unique string that can be used as an identifier when adding operations.
             if set to None, a random hash will be generated instead.
         Returns
@@ -687,7 +770,7 @@ class Schedule(ScheduleBase):  # pylint: disable=too-many-ancestors
 class CompiledSchedule(ScheduleBase):
     """
     A schedule that contains compiled instructions ready for execution using
-    the :class:`~.instrument_coordinator.InstrumentCoordinator`.
+    the :class:`~.InstrumentCoordinator`.
 
     The :class:`CompiledSchedule` differs from a :class:`.Schedule` in
     that it is considered immutable (no new operations or resources can be added), and
@@ -698,7 +781,7 @@ class CompiledSchedule(ScheduleBase):
         A :class:`~.CompiledSchedule` can be obtained by compiling a
         :class:`~.Schedule` using :func:`~quantify_scheduler.compilation.qcompile`.
 
-    """
+    """  # pylint: disable=line-too-long
 
     schema_filename = "schedule.json"
 
@@ -722,12 +805,12 @@ class CompiledSchedule(ScheduleBase):
         The contents of this dictionary depend on the backend it was compiled for.
         However, we assume that the general format consists of a dictionary in which
         the keys are instrument names corresponding to components added to a
-        :class:`~.instrument_coordinator.InstrumentCoordinator`, and the
+        :class:`~.InstrumentCoordinator`, and the
         values are the instructions for that component.
 
         These values typically contain a combination of sequence files, waveform
         definitions, and parameters to configure on the instrument.
-        """
+        """  # pylint: disable=line-too-long
         return self.data["compiled_instructions"]
 
     @classmethod