+++ tests/scheduler/helpers/test_inspect.py
@@ -6,9 +6,8 @@ from quantify_scheduler.helpers import inspect
 
 def test_get_classes():
     # Arrange
-    from quantify_scheduler import (
-        gate_library,
-    )  # pylint: disable=import-outside-toplevel
+    # pylint: disable=import-outside-toplevel
+    from quantify_scheduler import gate_library
 
     # Act
     classes = inspect.get_classes(gate_library)